% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_bracket.R
\name{add_bracket}
\alias{add_bracket}
\title{Add a Labelled Bracket to Group Predictors in a Dot-and-Whisker Plot}
\usage{
add_bracket(p, label, top, bottom, face = "italic")
}
\arguments{
\item{p}{A dot-and-whisker plot generated by \code{dwplot}}

\item{label}{The text to appear on the bracket}

\item{top}{The name of the topmost variable to be bracketed as a character string, or alternately, the numeric position of this variable, counting from the bottom of the y-axis}

\item{bottom}{The name of the bottommost variable to be bracketed as a character string, or alternately, the numeric position of this variable, counting from the bottom of the y-axis}

\item{face}{Options for label text: "plain", "bold", "italic", "oblique", and "bold.italic"}

\item{xloc}{}

\item{last}{logical: Is this the last bracket to be added to the plot?}
}
\value{
The function returns a \code{ggplot} (if \code{last = FALSE}) or \code{gtable} (if \code{last = TRUE}) object.
}
\description{
\code{add_bracket} draws a bracket along the y-axis beyond the plotting area of a dot-and-whisker plot generated by \code{dwplot}, useful for labelling a group of predictors
}
\examples{
data(mtcars)
m1 <- lm(mpg ~ wt + cyl + disp, data = mtcars)
m1_df <- tidy(m1) # create data.frame of regression results

p <- dwplot(m1_df) +
    scale_y_discrete(breaks = 4:1, labels=c("Intercept", "Weight", "Cylinders", "Displacement")) +
    theme_bw() + xlab("Coefficient") + ylab("") +
    geom_vline(xintercept = 0, colour = "grey50", linetype = 2) +
    theme(legend.position="none")

p \%>\% add_bracket(label="Engine", top="cyl", bottom=1) \%>\% add_bracket(label="Not Engine", top=4, bottom=3, xloc=-29, last = TRUE)
}

