% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{One or several ggplot objects separated by comma which are intended to combine.}

\item{plotlist}{A list of ggplot objects.}

\item{cols}{A numeric value setting the number of columns in the layout}

\item{A}{layout matrix. This setting will overwrite the setting in the \code{cols}.}
}
\value{
The function returns a unclassified plot file, which can be saved by regular plot devices, such as \code{pdf("<FileName>.pdf")...dev.off()}.
}
\description{
\code{multiplot} is a funciton for combining multiple ggplot objects into one page, offered by \href{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}{\emph{Cookbook for R}}.
}

