% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/secret_weapon.R
\name{secret_weapon}
\alias{secret_weapon}
\title{Format Regression Results for a 'Secret Weapon' Plot}
\usage{
secret_weapon(x, var = NULL, alpha = 0.05)
}
\arguments{
\item{x}{Either a tidy data.frame including results from multiple models (see 'Details') or a list of model objects that can be tidied with \code{\link[broom]{tidy}}}

\item{var}{The predictor whose results are to be shown in the 'secret weapon' plot}

\item{alpha}{A number setting the criterion of the confidence intervals. The default value is .05, corresponding to 95-percent confidence intervals.}
}
\description{
\code{secret_weapon} is a function for formatting regression results to be passed to \code{\link[dotwhisker]{dwplot}} to generate a 'secret weapon' plot
}
\details{
Andrew Gelman has coined the term \href{http://andrewgelman.com/2005/03/07/the_secret_weap/}{"the secret weapon"} for dot-and-whisker plots that compare the estimated coefficients for a single predictor across many models or datasets.
\code{secret_weapon} takes a tidy data.frame of regression results suitable for passing to \code{\link[dotwhisker]{dwplot}} or a list of model objects and reformats these results so that, when passed to \code{\link[dotwhisker]{dwplot}}, the result will be a dot-and-whisker plot of the results of a single variable across the multiple models.
}
\examples{
# Estimate models across many samples, put results in a tidy data.frame
by_clarity <- diamonds \%>\% group_by(clarity) \%>\%
 do(tidy(lm(price ~ carat + cut + color, data = .))) \%>\%
 ungroup \%>\% rename(model=clarity)

# Format for a 'secret weapon' plot of the results of diamond size, make the plot
secret_weapon(by_clarity, "carat") \%>\% dwplot
}

