% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smul.R
\name{smul}
\alias{smul}
\title{Create dataset list for small multiple plots}
\usage{
smul(df, by = "term", compare = NULL)
}
\arguments{
\item{df}{A data.frame including the variables \code{term} (names of independent variables), \code{estimate} (corresponding coefficient estimates), \code{std.error} (corresponding standard errors), and optionally \code{model} (when multiple models are desired on a single plot) such as generated those by \code{\link[broom]{broom::tidy}}.}

\item{by}{A character showing based on which variable the \code{df} is splited. For a typical "small multiple" plot, the spliting criterion is usually \code{df$term}.}

\item{compare}{A character indicating the new \code{model} in each splited group of estimates.}
}
\value{
The function returns a list of tidy datasets.
}
\description{
\code{smul} creates a list of datasets for ploting the "small multiple" plots (Kastellec and Leoni 2007, 768).
}
\examples{
data("diamonds")
library(broom)


diamonds$cut[diamonds$cut == "Fair"] <- "Good"

 by_cut <- diamonds \%>\% group_by(cut) \%>\%
  do(tidy(lm(price ~ depth + table, data = .))) \%>\% rename(model=cut)

by_cut$top <- 0
by_cut$top[by_cut$model \%in\% c("Premium", "Ideal")] <- 1

by_cut$model <- rep(c("low-level", "high-level"), each = 3, times = 2)
smul_list <- smul(by_cut, "term", "top")
}
\references{
Kastellec, Jonathan P. and Leoni, Eduardo L. 2007. "Using Graphs Instead of Tables in Political Science." Perspectives on Politics, 5(4):755-771.
}

