% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/by_2sd.R
\name{by_2sd}
\alias{by_2sd}
\title{Standardize results by multiplying by 2 standard deviations
Per:
Gelman, Andrew. 2008. Scaling regression inputs by dividing by two standard deviations.
     Statistics in Medicine, 27:2865–2873.}
\usage{
by_2sd(tr, df)
}
\description{
Standardize results by multiplying by 2 standard deviations
Per:
Gelman, Andrew. 2008. Scaling regression inputs by dividing by two standard deviations.
     Statistics in Medicine, 27:2865–2873.
}

