% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel_y_axis.R
\name{relabel_y_axis}
\alias{relabel_y_axis}
\title{Relabel the Y-Axis of a Dot-Whisker Plot}
\usage{
relabel_y_axis(x)
}
\arguments{
\item{x}{A vector of labels for predictors, listed from top to bottom}
}
\value{
The function returns an object of the same type as it is passed: a plot generated by \code{\link[dotwhisker]{dwplot}}.
}
\description{
\code{relabel_y_axis} DEPRECATED. A convenience function for relabeling the predictors on the y-axis of a dot-whisker plot created by \code{\link[dotwhisker]{dwplot}}.  It is deprecated; use \code{\link[dotwhisker]{relabel_predictors}} instead.
}
\seealso{
\code{\link[dotwhisker]{relabel_predictors}} to relabel predictors on the y-axis of a dot-whisker plot or in a tidy data.frame
}
