% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/small_multiple.R
\name{small_multiple}
\alias{small_multiple}
\title{Generate a 'Small Multiple' Plot of Regression Results}
\usage{
small_multiple(x, var = NULL, alpha = 0.05)
}
\arguments{
\item{x}{Either a tidy data.frame including results from multiple models (see 'Details') or a list of model objects that can be tidied with \code{\link[broom]{tidy}}}

\item{alpha}{A number setting the criterion of the confidence intervals. The default value is .05, corresponding to 95-percent confidence intervals.}
}
\value{
The function returns a \code{ggplot} object.
}
\description{
\code{small_multiple} is a function for plotting regression results of multiple models as a 'small multiple' plot
}
\details{
Kastellec and Leoni (2007)
\code{small_multiple} takes a tidy data.frame of regression results or a list of model objects and generates a dot-and-whisker plot of the results of a single variable across the multiple models.

Tidy data.frames to be plotted should include the variables \code{term} (names of predictors), \code{estimate} (corresponding estimates of coefficients or other quantities of interest), \code{std.error} (corresponding standard errors), and \code{model} (identifying the corresponding model).
In place of \code{std.error} one may substitute \code{lb} (the lower bounds of the confidence intervals of each estimate) and \code{ub} (the corresponding upper bounds).

Alternately, \code{small_multiple} accepts as input a list of model objects that can be tidied by \code{\link[broom]{tidy}}.
}
\examples{
# Estimate models across many samples, put results in a tidy data.frame
by_clarity <- diamonds \%>\% group_by(clarity) \%>\%
 do(tidy(lm(price ~ carat + cut + color, data = .))) \%>\%
 ungroup \%>\% rename(model=clarity)

# Format for a 'secret weapon' plot of the results of diamond size, make the plot
secret_weapon(by_clarity, "carat") \%>\% dwplot
}

